Fixing bug: Passing all required arguments to `findWaves`
Before, changing the reference period, type of wave or daily quantile threshold didn't have any effect on `findWaves` because this parameters were not passed to the corresponding functions. Now that is not an issue
NOTE: This change was checked and changed in all other function with the same problem. Also, some default values were updated.
<°))))><